// packages/cfdi/xsd/src/types/key-schema.ts
export enum Schemakey {
  TIPODATOS = 'CATALOGOS_TIPODATOS',
  CATALOGOS = 'CATALOGOS_CATALOGOS',
  COMPROBANTE = 'COMPROBANTE',
  INFORMACIONGLOBAL = 'COMPROBANTE_INFORMACIONGLOBAL',
  CFDIRELACIONADOS_CFDIRELACIONADO = 'COMPROBANTE_CFDIRELACIONADOS_CFDIRELACIONADO',
  CFDIRELACIONADOS = 'COMPROBANTE_CFDIRELACIONADOS',
  EMISOR = 'COMPROBANTE_EMISOR',
  RECEPTOR = 'COMPROBANTE_RECEPTOR',

  CONCEPTO_IMPUESTOS_TRASLADOS_TRASLADO = 'COMPROBANTE_CONCEPTOS_CONCEPTO_IMPUESTOS_TRASLADOS_TRASLADO',
  CONCEPTO_IMPUESTOS_TRASLADOS = 'COMPROBANTE_CONCEPTOS_CONCEPTO_IMPUESTOS_TRASLADOS',
  CONCEPTO_IMPUESTOS_RETENCIONES_RETENCION = 'COMPROBANTE_CONCEPTOS_CONCEPTO_IMPUESTOS_RETENCIONES_RETENCION',
  CONCEPTO_IMPUESTOS_RETENCIONES = 'COMPROBANTE_CONCEPTOS_CONCEPTO_IMPUESTOS_RETENCIONES',
  CONCEPTO_IMPUESTOS = 'COMPROBANTE_CONCEPTOS_CONCEPTO_IMPUESTOS',
  CONCEPTO_ACUENTATERCEROS = 'COMPROBANTE_CONCEPTOS_CONCEPTO_ACUENTATERCEROS',
  CONCEPTO_INFORMACIONADUANERA = 'COMPROBANTE_CONCEPTOS_CONCEPTO_INFORMACIONADUANERA',
  CONCEPTO_CUENTAPREDIAL = 'COMPROBANTE_CONCEPTOS_CONCEPTO_CUENTAPREDIAL',
  CONCEPTO_COMPLEMENTOCONCEPTO = 'COMPROBANTE_CONCEPTOS_CONCEPTO_COMPLEMENTOCONCEPTO',
  CONCEPTO_PARTE_INFORMACIONADUANERA = 'COMPROBANTE_CONCEPTOS_CONCEPTO_PARTE_INFORMACIONADUANERA',
  CONCEPTO_PARTE = 'COMPROBANTE_CONCEPTOS_CONCEPTO_PARTE',
  CONCEPTO = 'COMPROBANTE_CONCEPTOS_CONCEPTO',
  CONCEPTOS = 'COMPROBANTE_CONCEPTOS',

  IMPUESTOS_RETENCIONES_RETENCION = 'COMPROBANTE_IMPUESTOS_RETENCIONES_RETENCION',
  IMPUESTOS_RETENCIONES = 'COMPROBANTE_IMPUESTOS_RETENCIONES',
  IMPUESTOS_TRASLADOS_TRASLADO = 'COMPROBANTE_IMPUESTOS_TRASLADOS_TRASLADO',
  IMPUESTOS_TRASLADOS = 'COMPROBANTE_IMPUESTOS_TRASLADOS',
  IMPUESTOS = 'COMPROBANTE_IMPUESTOS',
  COMPLEMENTO = 'COMPROBANTE_COMPLEMENTO',
  ADDENDA = 'COMPROBANTE_ADDENDA',
}